Modified function of getting pivot
This modification changes the way of getting the pivot to median instead of
mean. This ensures more reliability in the correction of the project for
people that seek a specific case to fail it. The previous case was better in
the general case because getting the pivot was O(n) complexity, but was prone
to more moves with specific cases. The median has O(n^2) complexity but ensures
the moves stay constant. Median makes the overall program slower but the moves
stay low with any case.